move pysite and calcargs out from model
