+++ fract4d/c/model/site.h
@@ -5,7 +5,6 @@
 
 #include "model/enums.h"
 
-typedef struct _object PyObject;
 typedef struct s_pixel_stat pixel_stat_t;
 
 // a type which must be implemented by the user of
@@ -53,31 +52,6 @@ public:
     virtual void wait();
 protected:
     pthread_t tid;
-};
-
-// @TODO: this sub-class should be moved out of this model into the Python interface to keep this module portable
-class PySite : public IFractalSite
-{
-public:
-    PySite(PyObject *site_);
-    void iters_changed(int numiters);
-    void tolerance_changed(double tolerance);
-    void image_changed(int x1, int y1, int x2, int y2);
-    void progress_changed(float progress);
-    void stats_changed(pixel_stat_t &stats);
-    void status_changed(int status_val);
-    bool is_interrupted();
-    void pixel_changed(
-        const double *params, int maxIters, int nNoPeriodIters,
-        int x, int y, int aa,
-        double dist, int fate, int nIters,
-        int r, int g, int b, int a);
-    void interrupt();
-    void start();
-    ~PySite();
-private:
-    PyObject *site;
-    bool has_pixel_changed_method;
 };
 
 // write the callbacks to a file descriptor
